Theme code moved to coreUi module.
Jetpack compose bom removed - now there is just compose version specified to all packages. It was removed, because BOM caused fetching some compose libraries in newly added modules.
