+++ addThermometer/addThermometerPresentation/build.gradle.kts
@@ -19,7 +19,7 @@ android {
             isMinifyEnabled = false
             proguardFiles(
                 getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro",
+                "proguard-rules.pro"
             )
         }
     }
@@ -33,6 +33,7 @@ android {
 }
 
 dependencies {
+    implementation(project(":coreUi"))
 
     implementation(libs.androidx.core)
     implementation(libs.androidx.appcompat)
