+++ app/src/main/java/com/klewerro/mitemperature2alt/presentation/addThermometer/connecting/ThermometerConnectingScreen.kt
@@ -34,12 +34,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.klewerro.mitemperature2alt.R
+import com.klewerro.mitemperature2alt.coreUi.LocalSpacing
+import com.klewerro.mitemperature2alt.coreUi.theme.MiTemperature2AltTheme
 import com.klewerro.mitemperature2alt.presentation.addThermometer.ConnectThermometerEvent
 import com.klewerro.mitemperature2alt.presentation.addThermometer.ConnectThermometerState
 import com.klewerro.mitemperature2alt.presentation.addThermometer.ConnectThermometerViewModel
 import com.klewerro.mitemperature2alt.presentation.addThermometer.ConnectingStatus
-import com.klewerro.mitemperature2alt.ui.LocalSpacing
-import com.klewerro.mitemperature2alt.ui.theme.MiTemperature2AltTheme
 import kotlinx.coroutines.delay
 
 private const val SCREEN_CHANGE_DELAY = 1_500L
