+++ app/src/main/java/com/klewerro/mitemperature2alt/presentation/addThermometer/name/ConnectThermometerNameScreen.kt
@@ -24,14 +24,14 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.klewerro.mitemperature2alt.R
+import com.klewerro.mitemperature2alt.coreUi.LocalSpacing
+import com.klewerro.mitemperature2alt.coreUi.theme.MiTemperature2AltTheme
 import com.klewerro.mitemperature2alt.domain.model.ThermometerStatus
 import com.klewerro.mitemperature2alt.presentation.addThermometer.ConnectThermometerEvent
 import com.klewerro.mitemperature2alt.presentation.addThermometer.ConnectThermometerState
 import com.klewerro.mitemperature2alt.presentation.addThermometer.ConnectThermometerViewModel
 import com.klewerro.mitemperature2alt.presentation.util.UiText
 import com.klewerro.mitemperature2alt.presentation.util.clearFocusOnClick
-import com.klewerro.mitemperature2alt.ui.LocalSpacing
-import com.klewerro.mitemperature2alt.ui.theme.MiTemperature2AltTheme
 
 @Composable
 fun ConnectThermometerNameScreen(
