+++ app/src/main/java/com/klewerro/mitemperature2alt/presentation/addThermometer/search/components/DevicesList.kt
@@ -19,9 +19,9 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.res.stringResource
 import com.klewerro.mitemperature2alt.R
+import com.klewerro.mitemperature2alt.coreUi.LocalSpacing
 import com.klewerro.mitemperature2alt.domain.model.ScannedDeviceStatus
 import com.klewerro.mitemperature2alt.domain.model.ThermometerScanResult
-import com.klewerro.mitemperature2alt.ui.LocalSpacing
 
 @Composable
 fun DevicesList(
