+++ app/src/main/java/com/klewerro/mitemperature2alt/presentation/mainscreen/MainScreen.kt
@@ -13,10 +13,10 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.platform.LocalContext
+import com.klewerro.mitemperature2alt.coreUi.LocalSpacing
 import com.klewerro.mitemperature2alt.presentation.mainscreen.components.MainScreenThermometerBox
 import com.klewerro.mitemperature2alt.presentation.mainscreen.components.NoConnectedThermometersInformation
 import com.klewerro.mitemperature2alt.presentation.mainscreen.components.SavedThermometerBox
-import com.klewerro.mitemperature2alt.ui.LocalSpacing
 
 @Composable
 fun MainScreen(
