+++ app/src/main/java/com/klewerro/mitemperature2alt/presentation/mainscreen/TopBar.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.klewerro.mitemperature2alt.R
-import com.klewerro.mitemperature2alt.ui.LocalSpacing
+import com.klewerro.mitemperature2alt.coreUi.LocalSpacing
 
 @Composable
 fun TopBar(title: String, shouldBeButtonVisible: Boolean, onButtonClick: () -> Unit) {
