+++ core/src/main/java/com/klewerro/mitemperature2alt/core/MyClass.kt
@@ -1,4 +1,3 @@
 package com.klewerro.mitemperature2alt.core
 
-class MyClass {
-}
+class MyClass
