+++ gradle/libs.versions.toml
@@ -8,6 +8,7 @@ composeCompiler = "1.5.10"
 hilt = "2.51"
 ble-kotlin-library = "1.0.8"
 lifecycle = "2.7.0"
+compose = "1.6.7"
 coroutines = "1.8.0"
 junit5 = "5.9.3"
 room = "2.6.1"
@@ -18,18 +19,17 @@ rules = "1.5.0"
 androidx-core = { module = "androidx.core:core-ktx", version = "1.12.0" }
 androidx-lifecycle = { module = "androidx.lifecycle:lifecycle-runtime-ktx", version.ref = "lifecycle" }
 androidx-appcompat = { module = "androidx.appcompat:appcompat", version = "1.6.1" }
-compose-bom = { module = "androidx.compose:compose-bom", version = "2024.02.02"}
 compose-activity = { module = "androidx.activity:activity-compose", version = "1.8.2"}
-compose-ui = { module = "androidx.compose.ui:ui" }
-compose-ui-graphics = { module = "androidx.compose.ui:ui-graphics"}
-compose-ui-tooling = { module = "androidx.compose.ui:ui-tooling"}
-compose-ui-manifest = {module = "androidx.compose.ui:ui-test-manifest"}
-compose-ui-tooling-preview = { module = "androidx.compose.ui:ui-tooling-preview" }
-compose-material = { module = "androidx.compose.material:material" }
-compose-lifecycle-viewModel = { module = "androidx.lifecycle:lifecycle-viewmodel-compose" }
+compose-ui = { module = "androidx.compose.ui:ui", version.ref = "compose" }
+compose-ui-graphics = { module = "androidx.compose.ui:ui-graphics", version.ref = "compose" }
+compose-ui-tooling = { module = "androidx.compose.ui:ui-tooling", version.ref = "compose" }
+compose-ui-manifest = {module = "androidx.compose.ui:ui-test-manifest", version.ref = "compose" }
+compose-ui-tooling-preview = { module = "androidx.compose.ui:ui-tooling-preview", version.ref = "compose" }
+compose-material = { module = "androidx.compose.material:material", version.ref = "compose" }
+compose-lifecycle-viewModel = { module = "androidx.lifecycle:lifecycle-viewmodel-compose", version.ref = "lifecycle" }
 compose-lifecycle-runtime = { module = "androidx.lifecycle:lifecycle-runtime-compose", version.ref = "lifecycle" }
 compose-navigation = { module = "androidx.navigation:navigation-compose", version = "2.7.7" }
-compose-icons = { module = "androidx.compose.material:material-icons-extended" }
+compose-icons = { module = "androidx.compose.material:material-icons-extended", version.ref = "compose" }
 coroutines-core = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-core", version.ref = "coroutines" }
 timber = { module = "com.jakewharton.timber:timber", version = "5.0.1"}
 # hilt
@@ -78,7 +78,6 @@ androidX = [
     "androidx-appcompat"
 ]
 compose = [
-    "compose-bom",
     "compose-activity",
     "compose-ui",
     "compose-ui-graphics",
